Add {fix_urls,colorize,new}.py & SConstruct from https://github.com/nhoffman/borborygmi
The above Python files are needed by scons/SConstruct to build a site but
with the files in their current location (the repo https://github.com/nhoffman/borborygmi)
introduces a dependence on Noah's website.

This is isn't desirable, so move the Python files to org-export and modify SConstruct
according.

Now to create a site, ensure org-export is in the path, and move
Sconstruct-original (renamed to SConstruct) to the base of a Pelican/Static website.